Issue #1104 - added closure and reorganized unit tests
