+++ fhir-term/src/test/resources/META-INF/services/com.ibm.fhir.registry.spi.FHIRRegistryResourceProvider
@@ -1,1 +1,1 @@
-com.ibm.fhir.term.test.FHIRTermTestResourceProvider
+com.ibm.fhir.term.service.test.FHIRTermServiceTestResourceProvider
